backend/requirements.txt
fastapi
uvicorn
supabase
openai
python-multipart
python-dotenv